+++ tutor/templates/apps/openedx/settings/cms/development.py
@@ -2,6 +2,10 @@
 import os
 from cms.envs.devstack import *
 
+LMS_BASE = "{{ LMS_HOST }}:8000"
+LMS_ROOT_URL = "http://" + LMS_BASE
+FEATURES["PREVIEW_LMS_BASE"] = "preview." + LMS_BASE
+
 {% include "apps/openedx/settings/partials/common_cms.py" %}
 
 # Setup correct webpack configuration file for development