+++ tutor/templates/apps/openedx/settings/cms/production.py
@@ -7,9 +7,6 @@ from cms.envs.production import *
 ALLOWED_HOSTS = [
     ENV_TOKENS.get("CMS_BASE"),
     "cms",
-    "127.0.0.1",
-    "localhost",
-    "studio.localhost",
 ]
 
 {{ patch("openedx-cms-production-settings") }}