+++ tutor/templates/apps/openedx/settings/partials/common_cms.py
@@ -2,9 +2,12 @@
 
 ######## Common CMS settings
 
-
 STUDIO_NAME = u"{{ PLATFORM_NAME }} - Studio"
 MAX_ASSET_UPLOAD_FILE_SIZE_IN_MB = 100
+
+FRONTEND_LOGIN_URL = LMS_ROOT_URL + '/login'
+FRONTEND_LOGOUT_URL = LMS_ROOT_URL + '/logout'
+FRONTEND_REGISTER_URL = LMS_ROOT_URL + '/register'
 
 # Create folders if necessary
 for folder in [LOG_DIR, MEDIA_ROOT, STATIC_ROOT_BASE]: